src/fetch/standalone/requirements.txt
boto3
requests
pyyaml
hurry.filesize